converting to view architechture, writing interface, dockerfile
- restructured repo - removed view/ dir
- dockerfile added
- added flask to requirements (temporary)
- started testing orchestrator to view interface
