+++ .gitignore
@@ -97,3 +97,4 @@ chat_history.dat
 chat_history.dir
 LICENCE.txt
 LICENCE.txt
+/configs
