+++ requirements.txt
@@ -3,3 +3,4 @@ python-dotenv
 openai
 setuptools
 argparse
+flask
