+++ setup.py
@@ -1,8 +1,8 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="chatbot_view",
-    version="0.1",
+    name="streamlit_view",
+    version="1.0.0",
     packages=find_packages(),
     install_requires=[
         "streamlit",
